refactor: OutputInterface의 로직을 ReceiptRenderer로 이전 (#14)
