+++ src/main/java/lotto/Application.java
@@ -29,7 +29,8 @@ public class Application {
         InputConverter converter = new InputConverter(validator);
         InputInterface in = new InputInterface(Console::readLine, System.out::println, converter);
         NumberFormat numberFormat = NumberFormat.getIntegerInstance();
-        OutputInterface out = new OutputInterface(System.out::println, numberFormat);
+        ReceiptRenderer receiptRenderer = new ReceiptRenderer();
+        OutputInterface out = new OutputInterface(System.out::println, numberFormat, receiptRenderer);
 
         Controller controller = new Controller(in, out, username, purchaseLotteryService,
                 applyWinningLotteryService, calculateResultService);
